Reorganized shlr/gdb in preparation for gdbserver
